+++ lib/builder.dart
@@ -8,5 +8,6 @@ export 'package:lean_builder/src/builder/errors.dart';
 export 'package:lean_builder/src/build_script/annotations.dart';
 export 'package:lean_builder/src/build_script/parsed_builder_entry.dart';
 export 'package:lean_builder/src/asset/asset.dart';
+export 'package:lean_builder/src/asset/assets_reader.dart';
 export 'package:lean_builder/src/asset/errors.dart';
 export 'package:lean_builder/src/asset/package_file_resolver.dart' show PackageFileResolver;